+++ node/resolvers/search/index.ts
@@ -205,14 +205,9 @@ export const queries = {
 
         // const refId = (items[0]?.referenceId ?? []).find((ref: any) => ref.Key === 'RefId')?.Value ?? ''
         const { commertialOffer, sellerId, sellerName } = items[0].sellers[0]
-        const minQty =
-          (product['Minimum Order Quantity'] ?? []).find((d: string) => d) ??
-          '1'
 
         let availableQuantity = 0
-        let isAvailable = false
-        const unitMultiplier =
-          (items ?? []).find((item: any) => item)?.unitMultiplier ?? 1
+        let isAuthorized = false
 
         if (targetSystem.toUpperCase() === 'SAP') {
           const productPlants =
@@ -235,7 +230,7 @@ export const queries = {
               partialSum + current?.totalQuantity ?? 0,
             0
           )
-          isAvailable = selectedProductWearhouses.length > 0
+          isAuthorized = selectedProductWearhouses.length > 0
         } else if (targetSystem.toUpperCase() === 'JDE') {
           const { AvailableQuantity } = commertialOffer
 
@@ -247,7 +242,7 @@ export const queries = {
 
           availableQuantity =
             brandDataMatch?.trade === productBrand ? AvailableQuantity : 0
-          isAvailable = brandDataMatch?.trade === productBrand
+          isAuthorized = brandDataMatch?.trade === productBrand
         }
 
         const price = commertialOffer.SellingPrice
@@ -256,13 +251,40 @@ export const queries = {
           ? commertialOffer.Price
           : commertialOffer.ListPrice
 
-        const uom = (product['Unit of Measure'] ?? []).find(
-          (i: string) => i !== ''
+        const uomKey =
+          targetSystem.toUpperCase() === 'SAP'
+            ? 'Unit of Measure'
+            : 'JDE Unit of Measure'
+
+        const uomDescriptionKey =
+          targetSystem.toUpperCase() === 'SAP'
+            ? 'UOM_Description'
+            : 'JDE Unit of Measure Description'
+
+        const moqKey =
+          targetSystem.toUpperCase() === 'SAP'
+            ? 'Minimum Order Quantity'
+            : 'JDE Minimum Order Quantity'
+
+        const leadTimeKey =
+          targetSystem.toUpperCase() === 'SAP'
+            ? 'Material_Master_Lead_Time'
+            : 'JDE_Lead_Time'
+
+        const uom = (product[uomKey] ?? []).find((i: string) => i && i !== '')
+
+        const uomDescription = (product[uomDescriptionKey] ?? []).find(
+          (i: string) => i && i !== ''
         )
 
-        const uomDescription = (product.UOM_Description ?? []).find(
-          (i: string) => i !== ''
+        const moq =
+          (product[moqKey] ?? []).find((i: string) => i && i !== '') ?? '1'
+
+        const leadTime = (product[leadTimeKey] ?? []).find(
+          (i: string) => i && i !== ''
         )
+
+        const unitMultiplier = skuItem?.unitMultiplier ?? 1
 
         return {
           refid: skuRefId,
@@ -272,6 +294,8 @@ export const queries = {
           skuName: skuItem?.name,
           uom,
           uomDescription,
+          moq,
+          leadTime,
           linkText: product.linkText,
           price,
           availableQuantity,
@@ -279,9 +303,8 @@ export const queries = {
             id: sellerId,
             name: sellerName,
           },
-          availability: isAvailable ? 'available' : 'unavailable',
+          availability: isAuthorized ? 'authorized' : 'unauthorized',
           unitMultiplier,
-          minQty,
         }
       })
 
@@ -303,6 +326,8 @@ export const queries = {
           skuName: null,
           uom: null,
           uomDescription: null,
+          moq: null,
+          leadTime: null,
           linkText: null,
           price: null,
           availableQuantity: null,